+++ app/models/project_type.rb
@@ -1,6 +1,6 @@
-# Redmine plugin for xmera:isms called Project Types Plugin
+# Redmine plugin for xmera called Project Types Plugin.
 #
-# Copyright (C) 2017-18 Liane Hampe <liane.hampe@xmera.de>
+# Copyright (C) 2017-19 Liane Hampe <liane.hampe@xmera.de>.
 #
 # This program is free software; you can redistribute it and/or
 # modify it under the terms of the GNU General Public License
@@ -17,21 +17,20 @@
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
 class ProjectType < ActiveRecord::Base
-  unloadable
   include Redmine::SafeAttributes
   # Associated models
   has_many :projects_project_types
   has_many :projects, :through => :projects_project_types
   
   has_many :trackers, :through => :project_types_default_trackers
-  has_many :project_types_default_trackers, :dependent => :delete_all
-  
+  has_many :project_types_default_trackers, :dependent => :delete_all 
   has_many :project_types_default_modules, :dependent => :delete_all
+  has_many :issue_custom_fields, :through => :trackers
     
   # Validations
   validates_presence_of :name
   validates_uniqueness_of :name
-  
+   
   attr_protected :id
   acts_as_positioned