+++ app/overrides/custom_fields/form.rb
@@ -16,12 +16,27 @@
 # along with this program; if not, write to the Free Software
 # Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301, USA.
 
-class AddIdentifierColumnToProjectTypes < ActiveRecord::Migration
-  def self.up
-    add_column :project_types, :identifier, :boolean unless column_exists?(:project_types, :identifier)
-  end
-  
-  def self.down
-    remove_column :project_types, :identifier if column_exists?(:project_types, :identifier)
-  end
+# Target is redmines app/views/custom_fields/_form.html.erb file
+# Removes the list of projects from the custom_fields form since the the assignment
+# to a specific project is already defined by the assingment to a tracker which is 
+# still possible.
+
+unless Rails.env == 'test'
+
+Deface::Override.new(
+  virtual_path: 'custom_fields/_form',
+  name: 'disables-issue-related-lists',
+  remove: "fieldset.box",
+  namespaced: true,
+  sequence: {before: "disables-project-list"}
+)
+Deface::Override.new(
+  virtual_path: 'custom_fields/_form',
+  name: 'disables-project-list',
+  insert_after: "erb[silent]:contains('if @custom_field.is_a?(IssueCustomField)')",
+  partial: "custom_fields/issue_related_fields",
+  namespaced: true,
+  sequence: {after: "disables-issue-related-lists"}
+)
+
 end